requirements.txt
langchain
openai
faiss-cpu
requests
python-dotenv